Finalizer moved to hash::detail #10
