+++ include/nil/crypto3/hash/blake2b.hpp
@@ -14,6 +14,8 @@
 
 #include <nil/crypto3/hash/detail/haifa_construction.hpp>
 #include <nil/crypto3/hash/detail/haifa_stream_processor.hpp>
+#include <nil/crypto3/hash/detail/finalizer.hpp>
+
 
 namespace nil {
     namespace crypto3 {
@@ -103,7 +105,7 @@ namespace nil {
                         constexpr static const std::size_t digest_bits = policy_type::digest_bits;
                     };
 
-                    typedef ::nil::crypto3::detail::finalizer<typename params_type::digest_endian, typename policy_type::word_type, 
+                    typedef ::nil::crypto3::hash::detail::finalizer<typename params_type::digest_endian, typename policy_type::word_type, 
                                                               policy_type::word_bits, policy_type::block_bits> 
                         finalizer;
                     typedef haifa_construction<params_type, typename policy_type::iv_generator,
